refactor(asr): extract ASRState into _types.py
Move the ASRState dataclass out of core.py into a dedicated _types.py so
every variant (standard/adaptive/Juggler/Riemannian) can depend on the
fitted-state container without importing the calibration pipeline. core.py
re-imports it; public API and behavior are unchanged.
